requirements.txt
python-telegram-bot
openai
langchain-huggingface
langchain-chroma
chromadb
python-dotenv
sentence-transformers